add controllers get all users
