+++ controllers/backofficeControllers.js
@@ -1,0 +1,14 @@
+const { getAllForCamps } = require('../services/usersServices');
+
+const getAllUsersForCamps = async (req, res, next) => {
+  try {
+    const allUsers = await getAllForCamps();
+    res.json(allUsers);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = {
+  getAllUsersForCamps,
+};
