+++ controllers/userControllers.js
@@ -1,5 +1,4 @@
 const { decodeAuthToken } = require('../services/jwtServices');
-const { getAllForCamps } = require('../services/usersServices');
 
 const getUserData = async (req, res, next) => {
   try {
@@ -10,16 +9,6 @@ const getUserData = async (req, res, next) => {
   }
 };
 
-const getAllUsersForCamps = async (req, res, next) => {
-  try {
-    const allUsers = await getAllForCamps();
-    res.json(allUsers);
-  } catch (error) {
-    next(err);
-  }
-};
-
 module.exports = {
   getUserData,
-  getAllUsersForCamps,
 };
